hero_world/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "hero_world"
version = "0.1.0"
dependencies = [
    "pygame-ce",
    "pygame-gui",
]

[project.scripts]
hw = "hero_world.game:game"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
hero_world = ["*.json", "assets/**/*png"]
